Rearrage test utils
