+++ pkg/await/states/util.go
@@ -15,37 +15,8 @@
 package states
 
 import (
-	"fmt"
-
-	"github.com/pulumi/pulumi-kubernetes/pkg/await/recordings"
-	"github.com/pulumi/pulumi-kubernetes/pkg/clients"
-	"github.com/pulumi/pulumi-kubernetes/pkg/logging"
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-// mustCheckIfRecordingsReady runs the provider StateChecker against the recordings and returns true if the state
-// is Ready, false otherwise. The last set of Update messages is also returned. This function is intended to be
-// used with vetted test data and will panic on error.
-func mustCheckIfRecordingsReady(recordingPaths []string, checker *StateChecker) (bool, logging.Messages) {
-	var messages logging.Messages
-	for _, recordingPath := range recordingPaths {
-		records := recordings.MustLoadWorkflow(recordingPath)
-
-		for _, record := range records {
-			obj, err := clients.FromUnstructured(record)
-			if err != nil {
-				panic(fmt.Sprintf("FromUnstructured failed: %v", err))
-			}
-
-			messages = checker.Update(obj)
-			if checker.Ready() {
-				return true, messages
-			}
-		}
-	}
-
-	return false, messages
-}
 
 // fqName returns the fully qualified name of the object in the form `[namespace]/name`.
 // The namespace is omitted if it is "default" or "".
